+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - MenuData Scriptable Object
 - AbstractMenu component
 - SelectableSlider component
+- AudibleSlider component
 
 ### Changed
 - DelayedButton implements ISelectable, ISubmitable and IHighlightable interfaces